11.0
Major release

 - Added Orders radial menu (with limited use) when playing as Special Infected
 - Changed 1 line in Miscellaneous 4 radial menu
 - Added more responses to "Pounced other" line for L4D2 characters
 - Added custom talker scripts needed for new lines

This is now the first vocalizer which allows you to vocalize sounds when you're playing Versus or Scavenge as Special Infected.

Lines in Orders radial menu for Special Infected will work on local servers and dedicated servers (where this add-on is installed). They CANNOT WORK on official Valve servers!